Faux-Sound: Improve and Clean-Up
Signed-off-by: Shoaib0597 <Shoaib0595@gmail.com>

Originally, Faux Sound Control was developed by faux123. Further, Mr. AW (Aditya Wicaksono) ported this Sound Control to MSM8916 SoC.

However, the Port for our Source Lacked the Ability (because WCD-9320 Codec is not used in our Source) of an On/Off Switch (as done by flar2). Further, the Code was Dirty in the sense that there were many Lines of Codes commented-out, Registry-Size was wrong, etc.

After tinkering with this Driver for some time, I managed to get the above-mentioned things Sorted-Out.

Further, a Special Thanks to "robcore" for fixing the 'Wrong Default-Values at Boot' issue of the Driver.

I used Faux Sound Control v3.6 as base version and on top of it made my changes. Since I did 3 changes, I named this version v3.9 (v3.6 + 3 changes = v3.9). Now, the changelog of this version is as follows---
1. Fixed the Registry and Cache Size according to the Gains (MSM8916 supports fewer Gains).
2. Added an Enable/Disable Toggle with a Smart Workaround (this allows full compatibility with Kernel Adiutor-MOD).
3. Altered the Formatting of the Codes (looks cleaner and more beautiful now).